Handle editing actions in the WKWebView subclass
